made tests more strictly return true / false values. soon all logging will be redirected and it will only be pass/fail

+++ environments/ubuntu/test-server-init.sh
@@ -3,10 +3,16 @@ export DATABASE_ROOT="$(./get-config.sh projectRoot)/services/database"
 export ARCHIVER_ROOT="$(./get-config.sh projectRoot)/services/archiver"
 echo "Starting server tests"
 export PGPASSWORD=$(./get-config.sh databasePassword)
-psql -U central_library -d central_library -h localhost -f $DATABASE_ROOT/test-database.sql
+SQL_RESULT=$(psql -q -t -A -U central_library -d central_library -h localhost -f $DATABASE_ROOT/test-database.sql 2> /dev/null)
+if [[ "$SQL_RESULT" != "hello|world" ]]; then
+	exit 1
+fi
 export PGPASSWORD=""
-rm helloWorld
+rm $ARCHIVER_ROOT/helloWorld &> /dev/null
 g++-13 -std=c++23 $ARCHIVER_ROOT/helloWorld.cpp -o $ARCHIVER_ROOT/helloWorld
-$ARCHIVER_ROOT/helloWorld
+CPP_RESULT=$($ARCHIVER_ROOT/helloWorld)
+if [[ "$CPP_RESULT" != "202100" ]]; then
+        exit 1
+fi
 rm $ARCHIVER_ROOT/helloWorld
-echo "Ending server tests"
+echo "Server test succeeded"
